Speed up reload of graph for different project
Initially we tried to obtain a nicer transition from one graph to
another. Due to d3-graphviz bugs this usually caused errors in the
worker, which needed to be caught (via an onerror handler) and then
restart the whole load/parse by deleting the previous SVG.

By just removing the SVG and starting the parse from scratch, we can
ensure that the load is consistent and is still "smooth" due to the
spinner loading.